[Exprimental] Increase reboot timeout
The staging-test-with-rebase times out more often than desirable.

According to the docs reboot_timeout became available in Ansible 2.7,
and the default timeout value is 600.

/!\ This task seems to be shared by mutliple playbooks, increasing
the timeout globally is probably not the best way to address the issue.
Increasing the timeout at all may not be the best way to address
the issue.

See https://docs.ansible.com/ansible/latest/modules/reboot_module.html